Fix several bugs in Android and Linux * Fix invalid exponential moving average in Counter (Linux) * Make printing network switcher state tidy (Linux) * Add skipping verbose messages in some classes (Android)   - ClientAdapter   - MainActivity.ReceivingThread * Fix crash error in sending control message (Android) * Fix network switcher's state checking (Android) * Add handling EAGAIN in ServerSockets (Linux)   - RfcommServerSocket   - TcpServerSocket   - ServerSocket * Add handling kernel I/O buffer full (Linux) * Add 60s waiting after trace runner finishes (Linux)

+++ linux/inc/Core.h
@@ -209,11 +209,11 @@ public:
   static void receive_control_message_loop(ServerAdapter *adapter);
 
   /* Get statistics */
-  int get_ema_send_request_size() {
+  float get_ema_send_request_size() {
     return this->mSendRequestSize.get_em_average();
   }
 
-  int get_ema_send_arrival_time() {
+  float get_ema_send_arrival_time() {
     return this->mSendArrivalTime.get_em_average();
   }
